third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/open-close/close_noopener_beforeunload-expected.txt

This is a testharness.js-based test.
[FAIL] closing noopener window with 2 entries
  assert_equals: session history has multiple entries expected 2 but got 1
Harness: the test ran to completion.
